improve e2e structure
